LPS-58824 Move out to portal-cache bundle
